The /usr/sbin/rsct/bin/refrsrc command forces a refresh of the HMC/IVM and RMC status after a migration occurs.  We should do this from the drmgr command instead of assuming servicelog is installed.  A failure to invoke this command can cause additional migrations to be delayed for up to 10 minutes while the framework refreshes itself.
The current scenario has the refrsrc command invoked from a servicelog
notifier handler.  This works in most cases but will cause a delay on
systems that do not have servicelog installed.  Additionally, one can
build drmgr without servicelog support which would cause the notification
to never be sent to servicelog even if it is installed.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>