When making changes to increase the smt state, the cpus that are brought online come online with the same smt-snooze-delay they had when they were taken offline.  This leads to a mix of smt-snooze-delay states for a system.
This patch will update the cpus after an smt state change so that they all
have the same smt-snooze-delay value if that value was consistant across
the cpus prior to changing the smt state.  If there was a mix of
smt-snooze-delay values prior to the smt state change then no attempt is
made to update the smt-snooze-delay state.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>