The current logic in ppc64_cpu will loop through the maximum number of cpu threads (set to 1024) when trying to determine or set a particular cpu attribuite.
For example, when setting smt mode, the code would try to set cpu
thread 0..1024 either online or offline until a failure occurs.  This
usually was trying to set the smt mode for a cpu thread that does not exist.
While this is mostly harmless, the side effect was that an error was returned
even though the smt setting operation completed successfully.

This patch changes the logic so that it only tries to set cpu attributes for
the cpu threads that are present on the system.  This resolves the inadvertent
error return and is logically a bit nicer than trying all possible threads.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>