This patch is dependent on two additional patches that are queued for inclusion in the kernel.
http://lists.ozlabs.org/pipermail/linuxppc-dev/2009-November/078181.html
http://lkml.org/lkml/2009/11/26/368

With this new kernel functionality CPU  DLPAR can be done completely
in the kernel.  Under this new scheme the drmgr command only has to write
the drc-index of a cpu to /sys/devices/system/cpu/probe to have the cpu
dlpar added and hotplug added to the system.  The DLPAR remove of a cpu
involves writing the path in the device tree for the the cpu to be
removed to he /sys/sdevices/system/cpu/release file.  This will hotplug
remove the cpu and dlpar remove it.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>