Change drmgr to attempt to remove memory which may be held by a balloon driver.  This change ignores the 'removable' flag if AMS and CMM are active.
The memory balloon driver (CMM) for active memory sharing (AMS) makes use
of a new 'memory isolate kernel notifier' along with the 'memory hotplug
notifier' to free pages which are not migratable.  To make use of this,
drmgr must attempt to remove memory which would have not have the sysfs
'removable' flag set.

Signed-off-by: Robert Jennings <rcj@linux.vnet.ibm.com>