The get_bus_id function takes a location code and then searches /sys/bus/pci/slots/*/phy_location for a slot with the matching location code.  Without this patch the comparison only considers the characters up to the length of the input location which can cause a false match. For example, the following to would match.
Input location: U5877.001.00H4031-P1
False match:    U5877.001.00H4031-P10
                                   ^ Comparison stops here

The length of the input is 23 characters and the first 23 characters
are compared, this leaves off the last '0' on the string to compare and
erroneously returns the wrong bus.

In the case of a PHB remove operation, this can cause the incorrect slot
(and associated devices) to be powered off.

Signed-off-by: Robert Jennings <rcj@linux.vnet.ibm.com>