The snap man page indicates that the file /var/log/platform is collected.  This is not always true and can cause some confusion since the file does not exist on newer systems.  To avoid confusing users who might think the file should be collected, remove it from the man page listing of files.
The /var/log/platform file is where the rtas_errd daemon used to write
its data before the advent of servicelog.  On older systems that may
still be running an older rtas_errd daemon (diagela package) this
file would still exist so we should still try to collect it in the
snap command if its present.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>