For each attempt to add or remove a number of LMBs the search for an available LMB starts at the head of the lmb list.  When removing a number of LMBs on a system with AMS enabled (where the is_removable flag is ignored) this causes several issues.  First, the remove will attempt to remove an LMB already removed.  Secondly, the remove will attempt to remove LMBs which are not removable repeatedly in it's search which wastes a lot of time.
This patch changes get_available_lmb so that it takes an lmb pointer to
start it's search rather than the lmb_list.  Both add_lmbs and
remove_lmbs are changed to keep an lmb pointer as the starting point to
pass to get_available_lmb and this is incremented as lmbs are attempted.

Signed-off-by: Robert Jennings <rcj@linux.vnet.ibm.com>