The current implementation requires that user be root to use the ppc64_cpu command.  This can be an issue for some that want to use the ppc64_cpu command to read information.  Some of the information reported by ppc64_cpu is available to non-root users, and there are bits that could be made available to non-root users in the future with some kernel updates.
This patch updates the command to allow it to be run as a non-root user.  Instead
of enforcing that a user need to be root to run the command, the command now
uses the users permissions when accessing a file (either read or write) to determine
if the action is possible.

Signed-off-by: Nathan Fontenot <nfont@austin.ibm.com>