Convert drmgr messaging to use say().
The drmgr commands provides two methods for outputting messages, a dbg()
routine meant for debug messages and a err_msg() routine meant for
error messages.  It appears that over this distinction has lost some
of its meaning since both methods output to stderr.  Additionally,
it is past time to do some overhaul on the messages produced by drmgr to
better explain failures that occur.

This patch implements a first step by combining the dbg() and err_msg()
routines into a common routine named say().  The new routine takes an output
level of 1-4 corresponding to the levels of output that can be specified
with the -d option, plus a format string and optional args a la printf.

All of the output is converted to use the say() routine where former dbg()
statements are given a level 4 output and former err_msg() routines are given
level 1 output.  Except for one or two cases where I noticed that a dbg()
message should have been an err_msg() I have left the output exactly like it
was before. The goal of this patch is to make the first step in updating
messages by moving to a common output routine.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>