Remove drmgr drslot_* symlinks.
The drmgr command has had symlinks pointing to it as part of trying to
maintain backwards compatibility.  These are the /usr/sbin/drslot_*
links that are set up in the post install script of the spec file. At this
point these symlinks should not be needed (some distros currently do not
even create them on install) and can be done away with.

The symlinks were introduced when the various separate drslot_* commands
were merged into a single drmgr command to do away with duplicated code
across the commands.

Removing the symlinks also allows us to update some of the option checking
and reporting code.  This patch gets rid of some additional checking that
is no longer needed with the symlinks removed.  Additionally, the dr_init()
routine is called later to avoid doing work that is not needed when we
are just parsing and validating options.  This also avoids logging drmgr
invocations when a parameter is incorrect, the -h, or the -C options
are specified.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>