ppc64_cpu: Correct core onlining.
The current code allows for users to specify to online more cores than
are present on the system.  This should be considered and error. Instead
we now have situations where cores can be taken offline with the current
logic in the code.

This replaces the logic with a simple message to fail the request if the
number of cores requested is greater than the number of cores present in
the system.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>