The drmgr command has always taken the -w option to specify how long to try an operation before timing out. Historically this option was always ignored, and then at some point used to time out when trying to acquire the drmgr lock.
As we see larger systems coming around we need to start honoring
this option so we do not consume unneeded time going through every
possible LMB and CPU on the system.

This patch saves the timeout specified on the command line (-w)
and then checks it while trying to take the drmgr lock and before
each CPU/LMB it tries to add/remove. If no timeout is specified
then drmgr will never timeout.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>