drmgr: Ensure devices are present during pci replacement
In some instance during a PCI replacement operations (drmgr -c pci -R)
the adapter we are adding does not immediately appear. In these cases
a small wait is required for the adapter to appear.

This patch adds a small loop checking for the adapter's presence before
continuing the replace operation. The update also adds an option for
users to exit the operation if the adapter does not appear.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>