drmgr: Post processing for PCI replace
During a PCI replace operation we can hit scenarios where the adapter
was not able to completely clean everything up before it is removed.
Such a scenario would be an adapter completely failing or just being
pulled from its slot while the system is up.

In these cases we need to add an additional remove and add to the flow
of operations in the replace code path. This allows us to add the
adapter back to the system which gives us enough to do a complete
remove and clean everything up, and then add the adapter again.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>