Acquire the dr_lock earlier in dr operations.
The dr_lock used to serialize dr operations should be taken earlier.
The current code takes the lock after opening the log file, this can
lead to having multiple instances of drmgr writing to the file at the
same time.

To do this, the patch moves the taking and releasing of the dr_lock into
the dr_init and dr_fini routines as well as cleans up main() to make
unlocking easier to handle.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>