From: Philip Kelleher <pjk1939@linux.vnet.ibm.com>
Added support for IBM Flash Adapter devices within the
snap script.

The script will check whether or not IBM Flash
Adapters are currently loaded in the system. If so, the
script will check them. If not, the script will do its
normal operations.

It will, then, check whether or not the rsxx-utils are
installed on the system. If they are installed, the rs_cardreport
application will be ran. If not, the script will print out
a warning stating that IBM Flash Adapter information could
not be gathered. The script will continue normally either
way.

Signed-off-by: Philip J Kelleher <pjk1939@linux.vnet.ibm.com>