ppc64_cpu: Add --subcores-per-core option
Upcoming power8 chips support a concept called split core. This is where
the core can be split into subcores that although not full cores, are
able to appear as full cores to a guest.

Currently the kernel supports setting subcores per core to 1 or 4,
through a sysfs interface.

This command allows the split state to be queried, and set to 1 or 4.

Signed-off-by: Joel Stanley <joel@jms.id.au>