ppc64_cpu: Teach all options about subcores
Commands such as --smt=on/off need to know about subcores so they can
online the required threads when a core is in split core mode.

Instead of re-implementing commands so that subcores are treated
correctly, make the entire program treat subcores as cores. This means
that commands such as --cores-on report the number of subcores, not
physical cores.

Signed-off-by: Joel Stanley <joel@jms.id.au>