ppc64_cpu: Add --info command
This lists all threads on the system, grouped by the cores they belong
to. The online status of a thread is indicated by an asterisk.

  $ ppc64_cpu --info
  Core   0:    0*    1*    2*    3*    4*    5*    6*    7*
  Core   1:    8*    9*   10*   11*   12*   13*   14*   15*
  Core   2:   16*   17*   18*   19*   20*   21*   22*   23*
  Core   3:   24*   25*   26*   27*   28*   29*   30*   31*

If a system supports subcores, and the user has permission to access the
sysfs files that indicate the subcore status, the threads will be
grouped by subcore.

  $ ppc64_cpu --info
  Core   0:
    Subcore   0:    0*    1
    Subcore   1:    2     3
    Subcore   2:    4     5
    Subcore   3:    6     7
  Core   1:
    Subcore   4:    8*    9
    Subcore   5:   10    11
    Subcore   6:   12    13
    Subcore   7:   14    15

Signed-off-by: Joel Stanley <joel@jms.id.au>