ppc64_cpu: use write(2) when modifying sysfs
This improves the error handling when setting attributes fails. fprintf
does not return an error when a sysfs write returns an error.  Using the
write(2), we can catch bad return codes and inform the user.

The callers of set_attribute have been updated to ignore errors where
they are expected.

This fixes the following bug:

 Modifying --subcores-per-core when a guest is running is not allowed.
 sysfs reports -EBUSY but currently ppc64_cpu reports success, despite
 not doing anything.

Signed-off-by: Joel Stanley <joel@jms.id.au>