drmgr: endian fixes for ibm,configure-connector
The ibm,configure-connector RTAS call is supposed to be called several
times to do its job. It needs a work area to hold data between
subsequent calls. Among these data, there are integer values that
must be in big endian order when passed to RTAS.

A buffer is setup by drmgr and passed to the rtas_cfg_connector() call
in librtas for this purpose. The rtas_cfg_connector() call manipulates
the work area as simple pointer to a char buffer and makes no assumptions
about its content. It is hence not convenient to ask librtas to fix the
endianness of the work area.

Since the work area is entirely exposed in the drmgr code, it is far
easier and straightforward to handle endianess there. This patch
enforce the work area to be big endian, always, and drmgr to byteswap
when needed.

Signed-off-by: Greg Kurz <gkurz@linux.vnet.ibm.com>