drmgr: Update action option validation
The drmgr command does not do any validation of the action option
specified on the command line, except for some checking for CPU
operations.. The omission of an action option can lead to confusing
error messages or just exiting with an error.

This patch updates each connector type with a validation check for
their respective valid action options.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>