ppc64_cpu: update access validation routines
The use of fopen() to check for read/write access of system attribute
files (i.e. sysfs) no longer works with kernels 3.14 and later. The man
page for fopen() states that there is an implied file creation if the
file does not exist. This now causes fopen() to return EACCES instead
of ENOENT for files that do not exist.

This patch updates ppc64_cpu to use access() to validate file existence
and file read/write permissions.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Acked-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>