drmgr: Correct -s option handling
The usage statement for drmgr claims that we can add/remove memory and cpus
by drc name or drc index. The current code though assumes that any
use of the -s flag to specify this defaults to drc name.

This patch updates the option checking for memory and cpu operations to
allow users to specify a drc index with the -s option. This will be
handled the same way the -s option for pci devices is handled, assume it
is a drc name unless it starts with '0x', indicating the string is
really a hex value, and switching it to a drc index.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>