drmgr: do not remove the last CPU
The drmgr command should not attempt to remove the last cpu from the system.

Normally this would not be an issue as cpu remove operations are initiated
from the HMC and the HMC will not allow the last cpu to be removed. This
issue does arise when handling PRRN events and trying to update the affinity
on a partition that only has one cpu.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>