lparstat: using get_smt_mode when showing SMT info
get_smt_state is used in the system_data structure to display both the
shared_processor_mode and the smt_state members. After reviewing code and
documentation, it seems that the information in smt_state is meant to be
acquired from ppc64_cpu instead of lparcfg. With this change, the SMT listing
will be determined by the output of ppc64_cpu --smt.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>