drmgr: allow 'empty' state when disabling hp slot/adapter
Current code assumes that slots under a PHB are populated during PHB
removal. As a result, we attempt to unconfigure the slot/adapter by
powering it off, then checking for an 'unconfigured' state via
get_hp_adaptor_status/DR_ENTITY_SENSE. However, if the slot is empty
well get an 'emtpy' state instead.

We can avoid this by simply not attempting to unconfigure/poweroff an
adapter if the slot is empty, but it's possible we rely on this in
some cases to safely power-off slots. Instead, simply allow
DR_ENTITY_SENSE to report 'empty' after power-off in addition to
'unconfigured'.

Signed-by-off: Michael Roth <mdroth@linux.vnet.ibm.com>