drmgr: don't remove PHB's interrupt node unless PHB actually has one
Currently we fail to release a PHB to firmware if we're unable to remove
the PHB's interrupt-controller node from the guest device tree.

However, according to update_phb_ic_info():

"... there can be more ICs than PHBs on a system. In this case, some
ICs won't have my-drc-index."

In such cases, we don't assign the 'interrupt-controller*' path to the
PHB's OF node, so simply skip the IC node removal (it either doesn't
exist, or is not not owned by this PHB) when this occurs and proceed
with releasing the PHB.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>