drmgr: Correct memory affinity when adding memory
When adding a LMB we need to update the affinity as part of the add process.
The affinity for the LMB we are adding is conveyed in the
ibm,associativity property for the device tree node of the LMB we are adding.

To clarify, a little history. Previously LMB information in the device tree
was done as a node per LMB in the form memory@XXX, and all of the information
about the LMB as properties under this node. When we moved to using the
ibm,dynamic-reconfiguration-memory node to convey LMB information the
ibm,dynamic-memory property provides an index for each LMB into the
ibm,associativity-lookup-arrays property. This is where we determine the
affinity for each LMB.

In order to update the affinity for a LMB we are adding we have to call
configure connector to get the memory@XXX node and its properties for the
LMB we are adding (yes, rtas did not change this part), find the
ibm,associativity property, then find the entry in the associativity
lookup array that matches it. We then update the ibm,dynamic-memory
property with the proper index into the associativity lookup array.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Tested-by: Bharata B Rao <bharata@linux.vnet.ibm.com>