ppc64_cpu: Correct pthread handling for exiting threads
The pthread handling for checking cpu frequency in the ppc64_cpu
command does not properly handle 'soak' threads exiting from a failed
sched_setaffinity call.

We need to update the number of threads handled to be dynamically set
instead of defining MAX_NR_CPUS, which is out of date. We also need to
check for 'soak' threads that have exited before reading the perf counters

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>