drmgr: handle read failures in get_att_prop()
In get_att_prop(), though rarely, reading from /proc or /sysfs may
fail. This patch handles failures of these readings based on the fact,
if we successfully read the property, both fread() and fscanf() will
return 1. This patch also updates the fread() path so that we only
try to read the size of the file and not the size of the buffer
passed in.

This can also fix the compiler WARNING saying that 'rc' is unused
variable.

Signed-off-by: Boqun Feng <boqun.feng@linux.vnet.ibm.com>
Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>