drmgr: Fix file leak in dr_lock/dr_unlock interaction
This patch fixes a file leak in dr_lock(). The file descriptor
is meant to be saved to a global variable dr_lock_fd, but we
have an extraneous local variable named dr_lock_fd as well. So
When we get to dr_unlock, the global file descriptor that it
tries to close was never set and quietly fails.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>