errinjct: Initial contribution of errinjct tool
This patch contributes the previously IBM internal errinjct tool. The errinjct
tool is used to inject artificial hardware errors into a system. This version
of the code is functionally identical to the old internal version, but I have
made some superficial changes to remove IBM confidential information and
improve coding style.

Changes in v2: Now checks for platform. If a user is not on a PowerVM system,
the tool exits and prints a descriptive error message. Updated the man page to
document lack of support for non-PowerVM systems.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>