drmgr: Add missing free's of allocated memory
Using valgrind to do analysis of the memory usage of the drmgr command
revealed multiple instances of memory not being free'ed. Correct this
by adding appropriate calls to free commands.

Note that the analysis done, and this patch, does not cover all of the code
paths in drmgr. There may be additional patches as testing continues.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>