drmgr: Do not allow DLPAR of display adapters
The DLPAR of display adapters has never been supported but that support
has not ever been enforced in drmgr. Recent tests show that trying to DLPAR
remove a display adapter causes the drmngr command to appear hung as it
waits for display adapters to power down.

Add the ability to check for a drc-type of 'display' when trying to add
or remove a pci or slot, and verify that none of the children under a
PHB are display adapters when trying to add or remove a PHB.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>