lsslot/drmgr: Performance enhancement for parsing LMBs
The gathering of LMB information in the drmgr and lsslot commands is less
than optimal. The time taken to run the 'lsslot -c mem' command on a 32TB
system is roughly 30 minutes. The culprit is the way the commands create
a list of every lmb on the system then back through searching the list for
each of the LMBs when updating them later.

For systems using the ibm,dynamic-reconfiguration-memory device tree property
we can speed this up by updating the LMBs as we add them to the list. This
results in the performance increases sseen below for 'lsslot -c mem'.

Old Code:
real    31m2.407s
user    30m21.870s
sys     0m39.073s

New Code:
real    0m36.546s
user    0m1.235s
sys     0m17.724s

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>