ppc64_cpu: handle errors from strtol in --cores-on parsing
Currently, on a 24-core system, passing certain values to `ppc64_cpu
--cores-on` leads to some surprising results:

ppc64_cpu --cores-on=all
ppc64_cpu --cores-on=adfd
	offlines all but one core

This is because we are not using the endptr parameter of strtol to
determine if any of the string is consumed (and if none of the string is
consumed, strtol still returns a valid long value, namely 0).

ppc64_cpu --cores-on=12df
	onlines 12 cores

This is because we are not using the endptr parameter to strtol to
determine if the entire string is consumed (checking for trailing
characters).

Add both these checks.

After this patch:

ppc64_cpu --cores-on=all
Invalid number of cores to online: all
ppc64_cpu --cores-on=adfd
Invalid number of cores to online: adfd
ppc64_cpu --cores-on=12fd
Invalid number of cores to online: 12fd

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>