ppc64_cpu: add --{on,off}line-cores to {on,off}line specific cores
We currently have support for a mechanism to ensure an exact number of
cores (or all cores after a recent patch submission) are online.
However, there is no control over which cores are selected. Add two new
options to ppc64_cpu, --online-cores and --offline-cores to support
specifying which cores to bring online or offline, respectively. Both
options take either no parameters, in which case they output a
comma-separated list of cores that are online/offline, or a
comma-separated list of cores, in which case the command sets those
cores' states to online/offline.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>