ppc64_cpu: check return code from set_one_core
In both --cores-on and --online-cores, the return code from the helper
function set_one_core is not checked. This can lead to ppc64_cpu
returning 0 even though the command technically failed to do what was
requested. After this patch, --cores-on still doesn't return a proper
error code (fixing in a follow-on).

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>