drmgr: Correct LMB counting when discovering LMBs
Recent updates to how drmgr counted the number of LMBs found did
not save the count value correctly and in the case of LMBs found
in the ibm,dynamic-reconfiguration-memory property only counted
LMBs that were present. This caused havoc when trying to shuffle
the LMBs.

Save the LMB count in the lmb_list_head struct so we have it across
function calls. Also update the shuffle_lmbs to just take a pointer to
the lmb_list_head and make changes locally in the routine.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>