errinjct: Correct possible buffer overflow
In multiple places in the errinjct command we read the contents of
a file into a local buffer. This presents a possible issue if the
file we are reading ever exceeds the size of the buffer. To prevent
this issue a new read_file() routine is introduced that will
allocate a buffer large enough to hold the contents of the file and
read the file into the buffer.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>