rtas_dbg: Ensure we NULL terminate token names
The current code to look up rtas tokens writes the token name
to a static buffer size, this runs the risk of not having the
string null terminated if the string length matches the buffer
length.

Update the code to use strdup so we can ensure the strings are
null terminated. This also updates the free_rtas_tokens()
routine to free the token name.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>