drmgr: Load kernel modules for pci hotplug
For PHB DLPAR, slot DLPAR, and PCI hotplug operations we need to ensure
that the rpadlpar_io and rpaphp kernel modules are loaded or have been
built into the kernel. A check for this is currently done for PHB and
slot DLPAR when determining if the system is capable of these operations.

This patch moves the code to validate the presence of these modules
to be called from dr_init() for PHB/SLOT/PCI operations. In order
to do this the parsing of args has to be done prior to calling dr_init.
This should not cause an issue since we can parse args without holding
the dr_lock and any errors are written to stderr.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>