drmgr: Start using in-kernel DLPAR functionality for cpu/memory
Now that the capability to do DLPAR of cpus and memory can be
done completely in the kernel, update the drmgr command to use
this functionality.

The kernel updates provided a sysfs file (/sys/kernel/dlpar)
that will allow the drmgr command invoke DLPAR of cpus and memory
by writing to this file.

There are some exisitng behaviors of drmgr on the PowerVM platforms
that require cpu and memory to be done slightly differently. For
memory drmgr only needs to report the number of LMBs that were
added or removed so drmgr can just pass all requests straight to
the kernel. This patch also makes a common routine for reporting
memory resources to ensure the format is kept constant.

For cpu DLPAR the drmgr command has to report the drc name of each
cpu that is added/removed. To do this the command uses much of the
same device tree parsing that is currently used and make a request
to the kernel to remove a specific cpu (by drc index) where a
write to the sysfs probe file was previously done.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>