drmgr: Workaround for concurrent PRRN collisions
Proposed workaround for issues with current PRRN event handling mechanism.

Currently, the mechanism works by kicking off a script that calls a drmgr
remove followed by a drmgr add for each effected device. With this method,
problems can occur when multiple prrn events arrive in a short time span or
when other dlpar operations are happening while the script is running. What
can happen is that these other dlpar operations can jump in inbetween when
the script calls the remove and the add and take the drmgr lock. Then when
we try to add the memory back from the script, we are unable to take the
drmgr lock and the add fails.

This patch fixes this issue by creating a new -P flag for drmgr which will
perform the necessary removes/adds from within the drmgr lock. The new
option will be invoked as follows:

drmgr -P <path to prrn log file>

On the rtas_errd side, we will begin by creating a log file for the
PRRN event which will contain a list of the type and the drc index for
each effected device. Then instead of invoking the script, we will call
the above drmgr command with the path to the prrn log file. Then, on the
drmgr side, we parse the list and perform the necessary removes/adds and
then delete the prrn log file.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>