errinjct: Correct read_file() error checking
When reading files from syfs the file length returned from a stat()
call does not always match the actual length of the file. The read_file()
routine should be checking for an error return from the read() call
instead of checking that the amount read matches the file length returned
from a stat() call.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>