lparstat: flush data to stdout
The stdout stream is not being flushed after being written to.  This
means that it doesn't behave properly in shell pipelines as the data
isn't visible until the program exits.  The fix is to simply flush
after writes so the data is immediately available to other processes
in the pipeline.

Signed-off-by: Ian Neal <ian.neal@gmail.com>