drmgr: Disable use of in-kernel cpu hotplug
After submitting the patch to update drmgr to use the in-kernel hotplug
capabilities for memory and cpu, commit <bbc5fc07c54d> (drmgr: Start using
in-kernel DLPAR functionality for cpu/memory), I realized there
are still distro kernels that do not have the in-kernel cpu hotplug
capabilities. Using the latest drmgr with this update breaks cpu dlpar
because the cpu dlpar code assumes that if the dlpar sysfs file exists
then in-kernel cpu dlpar is possible. this is wrong, the sysfs dlpar file
was added with in-kernel memory hotplug, its existence can only guarantee
memory hotplug capability.

For now, disable using the in-kernel cpu capabilities by drmgr until
this can be resolved.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>