lparstat: Fix computation of time used in statistics
The routine get_time computes bogus timestap values due to confusing
seconds and microseconds.  This leads to incorrect return values from
elapsed_time and therefore incorrect values for the physc field.

Fixed by converting the "time" sysentry field to microseconds (and
computing it correctly), and updating the user of elapsed_time.

In addition, %entc is supposed to show the ratio of physc to the
number of entitled CPUs as *percentage*, so the ratio needs to be
multiplied by 100.

Signed-off-by: Ulrich Weigand <ulrich.weigand@de.ibm.com>
Signed-off-by: Michael Bringmann <mbringm@us.ibm.com>