lparstat: Fix bad cast from integer to float
Fix a bug where the cpu stats (%user, %sys, %wait, and %idle) do not add
up to 100%. The problem was that we were casting the cpu stats and the
total cpu time to floats from ints. For large integer values, this would
cause the floating point representation to be truncated. This solution
avoids the casts and stores the cpu stats as long longs and only casts
'total' to a long double in order to perform floating point division.

Additionally, this patch parses the proc stat values as long longs as
on systems that have been running for a long time, these values can exceed
the maximum 32 bit integer limit.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>