drmgr: Make option struct member 'action' global
Move the action member out of struct options to become the global
usr_action. Also introduce the enum drmgr_action as the type for
usr_action.

This patch also introduces the file options.c which will hold the
definitions of all of the global variables common to the drmgr and
lsslot commands. This file is included in drmgr.c and lsslot.c instead
of being built directly.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>