lparstat: Fix segfault when parsing /proc/interrupts
Fix a bug where we hit a segfault running 'lparstat -i' on machines with a
large number of cpus. The 'SPU' line in proc interrupts would exceed the
512B allocated and we would subsequently walk off the end of the array
when we attempt to parse the line. This solution uses getline instead of
fgets in order to easily handle lines of arbitrary length.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>