drmgr: Add options.c to Makefile for creating tarballs
Previous updates that ctreated the file options.c that is #included
instead of built directly for drmgr and lsslot did not update the Makefile
to ensure this file is included in tarballs when doing a 'make dist'.

Add options.c to the list of NOINST_HEADERS for drmgr and lsslot.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>