lparstat: Ignore whitespace at beginning of /proc/interrupts SPU line
In the case that any interrupt number in /proc/interrupts exceeds 3
characters, whitespace will be placed at the beginning of the SPU line in
order to keep the columns aligned. In that case, the current code will miss
the SPU line. This patch skips any whitespace preceding the SPU line.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>