drmgr: Add capability for PCI hotplug only operations
We need to provide the capability to do a hotplug only style of operation
for some adapters such as nvme adapters behind a bridge that can be owned
by more than one partition. These scenarios are used for doing failovers
and this update reduces the total drmgr time from 18 seconds per add or
remove down to 2-3 seconds per action.

Note that this is different from the PAPR notion of hotplug where the
device is hotplug removed from the OS, then the slot is powered down,
a new card inserted, the slot powered up, and the device hootplug added
back into the system. This is the drmgr -R action. This new functionality
just does the hotplug to/from the OS, nothing else.

This update assumes that when doing an add, the DR indicator already has
the proper isolation and allocation state, and the power to the slot is
on. For remove operations the indicator and power for the slot is not
modified.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>