ppc64_cpu: add a numeric option when printing SMT
Currently, there is not a numeric (non verbose) display of SMT modes,
which means that three verbose output is usually generated, as:

  SMT=X
  SMT is off
  Machine is not SMT capable

This patch implements a easily machine parsable output for SMT, helping
a lot of scripts that needs to read SMT status on a machine.

I am considering that if the SMT is off or impossible, then SMT=1 (1
thread per core), thus, always returning SMT=X, where X means thread
enabled per core, as 1, 2, 4, 8.

Signed-off-by: Breno Leitao <breno.leitao@gmail.com>