lsslot: Validate cpu->name when printing
When printing the information for all possible CPUs we can possibly print
a NULL string pointer. the libc printf code handles this gracefully by
printing '(NULL)' but this is not what we should be relying on.

This validates the cpu->name pointer before printing it and instead
prints '--' if cpu->name is not valid.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>