ppc64_cpu: Treat CPUs as online if cpu hotplug not enabled
cpu_online() looks for the "online" attribute to decide if a cpu is
online, however if the kernel does not have CONFIG_HOTPLUG_CPU enabled
the attribute will not exist.
If hotplug is not enabled then the cpu can be assumed to be online - in
this case, update cpu_online() to treat the cpu as online.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>