Improve perf of drmgr/lsslot with large num of virt. dev.
The current algorithm for add_pci_vio_node walks the vdevice bus
for every theoretical DRC (1000's of times).  The more vdevices
you have populated, the worse this algorithm becomes.  On a system
with 450 devices, this can take over 30 seconds for example.

This modifies the algorithm to populate a node list of existing
devices (so walks the tree just once), and then uses these nodes
when populating the full list instead of walking the device tree
for each one.

NOTE:  This only is changed for virtual devices - a system with
large numbers of PCI devices will not be optimized, but that is
less of a concern.

Signed-off-by: Amartey Pearson <apearson@us.ibm.com>
Signed-off-by: Bryant G. Ly <bryantly@linux.vnet.ibm.com>
[tyreld: changed my_drc_index type from u32 to uint32_t]
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>