drmgr: Stale errno usage corrections
Several places in the drmgr code use a stale value of errno either
as a return code or when printing an error message. A common example
of this is calling the say() routine and then returning the value of
errno. The say() routine makes several system calls and the value of
errno after a call to say() no longer reflects the original errno
value.

Several places in the drmgr code were updated to properly handle
the return codes from functions such as write() or open() where we
should be checking for rc == -1 instead of rc <= 0. This does make
a difference as errno won't be set for rc == 0.

Note that partial write return codes are not really handled in the
drmgr code. The drmgr command is writing to sysfs and /proc
files where we know that they will either return success or
failure and set errno, no partial writes.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Reviewed-by: Paul Clarke <pc@us.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>