lsslot: Validate DRC property groups
Recently seen on a kvm guest running on a PowerNV system, the
lsslot command segfaulted when trying to free the drc information.

# gdb -q ./lsslot
Reading symbols from ./lsslot...done.
(gdb) run -c memory
Starting program: ./lsslot -c memory

Program received signal SIGSEGV, Segmentation fault.
free_drc_info () at src/drmgr/common_ofdt.c:287
287			all_drc_lists = list->all_next;

The root cause of this is that the DRC properties in the guest were empty.
This caused the lsslot (and likely possible in drmgr since this code
is common to both commands) to improperly parse the device tree
information.

To fix this we need to add a check after gathering information for
each of the DRC properties when reading DRC property group data. If
there are no entries in any of the DRC properties we cannot continue.

Reported-by: Seeteena Thoufeek <s1seetee@linux.vnet.ibm.com>
Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>