ppc64_cpu: Treat CPUs as online if cpu hotplug not enabled
cpu_online() looks for the "online" attribute to decide if a cpu is
online, however if the kernel does not have CONFIG_HOTPLUG_CPU enabled
the attribute will not exist.
get_attribute() calls access() first to check if the attribute exists,
which will set errno with ENOENT if it is missing. If get_attribute()
fails with this specific error, treat the cpu as online.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>
[tyreld: removed unnecessary braces from if block]
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>