ppc64_cpu: Scale CPU cycles counter by amount of time it ran
At the moment we start a CPU cycles counter, sleep one second and
read the counter value, assuming we got one second of time on the
PMU.

There are two issues with this. If someone else is consuming all
the available PMCs, we may end up context switching the PMC and
count for half (or less) of the time. In this case we report a CPU
frequency that is half (or less) of what it should be.

On the other hand if the sleep goes for longer than one second which
might happen on a busy box, then we count for too long and report a
CPU frequency that is too high.

Also check that we got a reasonable number of cycles and report
an error if we did not.

Signed-off-by: Anton Blanchard <anton@samba.org>