lparstat: Fix parsing of /proc/interrupts SPU line
The current means of parsing the SPU line in /proc/interrupts is incorrect.
Adding 11 to the current pointer is only correct if the number of SPU
interrupts in each column is a single digit. Otherwise, for values with
more than a single digit, we will only see the last digit of the value.
This patch fixes the issue by using strstr to check for the SPU line and
strtok to split the line into tokens in order to omit the arbitrary
whitespace between values.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Reported-by: Flos Lonicerae <lonicerae@gmail.com>
Reviewed-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>