Test for librtas/librtasevent headers and libraries
The current configure script only tests for the presence of librtas.h. Certain
distros have chosen to split the librtas package into seperate librtas and
librtasevent library/development packages. As a result compilation will fail
for the rtas_event_decode utility if librtasevent.h is missing.

Fix this by also checking for librtasevent.h, and for sanity purposes that
the libraries themselves are actually present and linkable.

Reported-by: Anton Blanchard <anton.blanchard@gmail.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>