lsslot: add DRC Type description strings for new U.2 PCIe slots
Two new PCI slot types can be reported from firmware in the DRC Type
property for a PCI device. The following are the slot descriptions for
DRC types 43 and 44.

43	A U.2 PCI Express Rev 3 4x lane slot with 4x lanes connected
44	A U.2 PCI Express Rev 4 4x lane slot with 4x lanes connected

Currently, the lsslot utility reports "Unknown" for the description of
these pci devices. Update the php_slot_type_msg[] with these new slot
descriptions so that lsslot correctly reports them.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>