lsslot: Add ibm,dynamic-memory-v2 parsing capability
In order to support reporting memory LMB information the lsslot
command needs to know how to parse the ibm,dynamic-memory-v2
property. This patch adds this support to the shared
drslot_chrp_mem.c file.

In order to prevent the drmgr command from parsing this v2 property
a new flag is introduce, is_lsslot_cmd. The reason for this is that
we do not support memory DLPAR for ibm,dynamic-memory-v2 from
userspace.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>