lparstat: Show available physical processors in the shared pool
This patch adds the "app" field that displays the available
processors in the shared pool.

The platform provides the summation of the physical processor
pool idle cycles through H_PIC hypercall. This data is updated
in the "pool_idle_time" field of the /proc/ppc64/lparcfg file.
This patch derives the "app" from the "pool_idle_time" field.

v2: Handle fgets() errors

Signed-off-by: Aravinda Prasad <aravinda@linux.vnet.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>